+++ src/main/java/org/prgrms/wumo/domain/member/repository/RefreshTokenRepository.java
@@ -1,8 +1,8 @@
-package org.prgrms.wumo.global.util;
+package org.prgrms.wumo.domain.member.repository;
 
 import java.time.Duration;
 
-import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Repository;
 
@@ -12,19 +12,19 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class RefreshTokenRepository {
 
-	private final RedisTemplate redisTemplate;
+	private final StringRedisTemplate stringRedisTemplate;
 
 	public void save(String key, String value, long expireSeconds) {
-		ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
+		ValueOperations<String, String> valueOperations = stringRedisTemplate.opsForValue();
 		valueOperations.set(key, value, Duration.ofSeconds(expireSeconds));
 	}
 
 	public String get(String key) {
-		ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
+		ValueOperations<String, String> valueOperations = stringRedisTemplate.opsForValue();
 		return valueOperations.get(key);
 	}
 
 	public void delete(String key) {
-		redisTemplate.delete(key);
+		stringRedisTemplate.delete(key);
 	}
 }
