+++ src/main/resources/db/migration/V11__alter_member_drop_refresh_token.sql
